% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinearDetect-package.R
\name{constant.sim.break}
\alias{constant.sim.break}
\title{Generate the constant model data with break points}
\usage{
constant.sim.break(nobs, cnst, sigma, brk = nobs + 1)
}
\arguments{
\item{nobs}{number of time points}

\item{cnst}{the constant}

\item{sigma}{covariance matrix of the white noise}

\item{brk}{vector of break points}
}
\value{
A list object, which contains the followings
\describe{
  \item{series_y}{matrix of response data}
  \item{noises}{matrix of white noise error}
}
}
\description{
Generate the constant model data with break points
}
