% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinearDetect-package.R
\name{ggm.sim.break}
\alias{ggm.sim.break}
\title{Generate the gaussian graphical model data with break points}
\usage{
ggm.sim.break(nobs, px, sigma, brk = nobs + 1)
}
\arguments{
\item{nobs}{number of time points}

\item{px}{the number of features}

\item{sigma}{covariance matrix of the X matrix}

\item{brk}{vector of break points}
}
\value{
A list object, which contains the followings
\describe{
  \item{series_x}{matrix of data}
}
}
\description{
Generate the gaussian graphical model data with break points
}
