% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lambda_warm_up_lm}
\alias{lambda_warm_up_lm}
\title{lambda warm up for linear regression model}
\usage{
lambda_warm_up_lm(data_y, data_x, blocks, cv_index)
}
\arguments{
\item{data_y}{input matrix Y}

\item{data_x}{input matrix X}

\item{blocks}{the vector of blocks}

\item{cv_index}{the vector of indices for validation}
}
\value{
a value for parameter lambda
}
\description{
lambda warm up for linear regression model
}
