% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBFLChangePointDetection-package.R
\name{pred.block}
\alias{pred.block}
\title{Prediction function (block)}
\usage{
pred.block(X, phi, j, p.x, p.y, h)
}
\arguments{
\item{X}{data for prediction}

\item{phi}{parameter matrix}

\item{j}{the start time point for prediction}

\item{p.x}{the dimension of data X}

\item{p.y}{the dimension of data Y}

\item{h}{the length of observation to predict}
}
\description{
Prediction function (block)
}
