% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinearDetect-package.R
\name{pred.block.var}
\alias{pred.block.var}
\title{Prediction function for VAR (block)}
\usage{
pred.block.var(Y, phi, q, TT, p, h)
}
\arguments{
\item{Y}{data for prediction}

\item{phi}{parameter matrix}

\item{q}{the AR order}

\item{TT}{the start time point for prediction}

\item{p}{the number of time series components}

\item{h}{the length of observation to predict}
}
\value{
prediction matrix
}
\description{
Prediction function for VAR (block)
}
