% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinearDetect-package.R
\name{pred}
\alias{pred}
\title{prediction function}
\usage{
pred(X, phi, j, p.x, p.y, h = 1)
}
\arguments{
\item{X}{data for prediction}

\item{phi}{parameter matrix}

\item{j}{the start time point for prediction}

\item{p.x}{the dimension of data X}

\item{p.y}{the dimension of data Y}

\item{h}{the length of observation to predict}
}
\value{
prediction matrix
}
\description{
prediction function
}
