% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBFLChangePointDetection-package.R
\name{pred.var}
\alias{pred.var}
\title{Prediction function for VAR 2}
\usage{
pred.var(Y, phi, q, T, p, h = 1)
}
\arguments{
\item{Y}{data for prediction}

\item{phi}{parameter matrix}

\item{q}{the AR order}

\item{T}{the start time point for prediction}

\item{p}{the number of time series components}

\item{h}{the length of observation to predict}
}
\description{
Prediction function for VAR 2
}
