% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TBFLChangePointDetection-package.R
\name{remove.extra.pts}
\alias{remove.extra.pts}
\title{helper function for detection check}
\usage{
remove.extra.pts(pts, brk)
}
\arguments{
\item{pts}{the estimated change points}

\item{brk}{the true change points}
}
\description{
helper function for detection check
}
