% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{soft_full}
\alias{soft_full}
\title{soft threshold function}
\usage{
soft_full(L, lambda)
}
\arguments{
\item{L}{input matrix}

\item{lambda}{threshold parameter}
}
\value{
thresholded matrix L
}
\description{
soft threshold function
}
