% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinearDetect-package.R
\name{var.sim.break}
\alias{var.sim.break}
\title{Generating non-stationary ARMA data.}
\usage{
var.sim.break(
  nobs,
  arlags = NULL,
  malags = NULL,
  cnst = NULL,
  phi = NULL,
  theta = NULL,
  skip = 200,
  sigma,
  brk = nobs + 1
)
}
\arguments{
\item{nobs}{number of time points}

\item{arlags}{the true AR order}

\item{malags}{the true MA order}

\item{cnst}{the constant}

\item{phi}{parameter matrix of the AR model}

\item{theta}{parameter matrix of the MA model}

\item{skip}{the number of time points to skip at the begining (for stable data)}

\item{sigma}{covariance matrix of the white noise}

\item{brk}{vector of break points}
}
\value{
Matrice of time series data and white noise data
}
\description{
Generating non-stationary ARMA data.
}
