% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinearDetect-package.R
\name{BIC}
\alias{BIC}
\title{BIC  and HBIC function}
\usage{
BIC(residual, phi, gamma.val = 1, method = "MLR")
}
\arguments{
\item{residual}{residual matrix}

\item{phi}{estimated coefficient matrix of the model}

\item{gamma.val}{hyperparameter for HBIC, if HBIC == TRUE.}

\item{method}{method name for the model: MLR: Multiple Linear Regression; VAR: Vector autoregression;}
}
\value{
A list object, which contains the followings
\describe{
  \item{BIC}{BIC value}
  \item{HBIC}{HBIC value}
}
}
\description{
BIC  and HBIC function
}
