% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinearDetect-package.R
\name{mspe.plot}
\alias{mspe.plot}
\title{Plot the cross-validation score}
\usage{
mspe.plot(pred.error, lambda)
}
\arguments{
\item{pred.error}{prediction error}

\item{lambda}{indice of tuning parameter lambda}
}
\value{
No return value, called for plot
}
\description{
Plot the cross-validation score
}
